moved tests to latest
