+++ src/test/java/net/snowflake/client/jdbc/ConnectionIT.java
@@ -645,34 +645,6 @@ public class ConnectionIT extends BaseJDBCTest {
   }
 
   /**
-   * Verify the passed heartbeat frequency, which is too small, is changed to the smallest valid
-   * value.
-   */
-  @Test
-  public void testHeartbeatFrequencyTooSmall() throws Exception {
-    Properties paramProperties = new Properties();
-    paramProperties.put(CLIENT_SESSION_KEEP_ALIVE_HEARTBEAT_FREQUENCY, 2);
-    Connection connection = getConnection(paramProperties);
-
-    connection.getClientInfo(CLIENT_SESSION_KEEP_ALIVE_HEARTBEAT_FREQUENCY);
-
-    for (Enumeration<?> enums = paramProperties.propertyNames(); enums.hasMoreElements(); ) {
-      String key = (String) enums.nextElement();
-      ResultSet rs =
-          connection
-              .createStatement()
-              .executeQuery(String.format("show parameters like '%s'", key));
-      rs.next();
-      String value = rs.getString("value");
-
-      assertThat(key, value, equalTo("900"));
-    }
-
-    SFSession session = connection.unwrap(SnowflakeConnectionV1.class).getSfSession();
-    assertEquals(900, session.getHeartbeatFrequency());
-  }
-
-  /**
    * Verify the passed heartbeat frequency, which is too large, is changed to the maximum valid
    * value.
    */
@@ -697,33 +669,6 @@ public class ConnectionIT extends BaseJDBCTest {
     }
     SFSession session = connection.unwrap(SnowflakeConnectionV1.class).getSfSession();
     assertEquals(3600, session.getHeartbeatFrequency());
-  }
-
-  /**
-   * Verify the passed heartbeat frequency matches the output value if the input is valid (between
-   * 900 and 3600).
-   */
-  @Test
-  public void testHeartbeatFrequencyValidValue() throws Exception {
-    Properties paramProperties = new Properties();
-    paramProperties.put(CLIENT_SESSION_KEEP_ALIVE_HEARTBEAT_FREQUENCY, 1800);
-    Connection connection = getConnection(paramProperties);
-
-    connection.getClientInfo(CLIENT_SESSION_KEEP_ALIVE_HEARTBEAT_FREQUENCY);
-
-    for (Enumeration<?> enums = paramProperties.propertyNames(); enums.hasMoreElements(); ) {
-      String key = (String) enums.nextElement();
-      ResultSet rs =
-          connection
-              .createStatement()
-              .executeQuery(String.format("show parameters like '%s'", key));
-      rs.next();
-      String value = rs.getString("value");
-
-      assertThat(key, value, equalTo(paramProperties.get(key).toString()));
-    }
-    SFSession session = connection.unwrap(SnowflakeConnectionV1.class).getSfSession();
-    assertEquals(1800, session.getHeartbeatFrequency());
   }
 
   @Test
